+++ static/scripts/pwd.js
@@ -5,11 +5,15 @@ $(document).ready(() => {
 		const passwordId = $($(event.target).parents('div')[1]).find('input').attr('id');
 
 		if ($(`#${passwordId}`).attr('type') === 'text') {
-			$(`#${passwordId}`).attr('type', 'password');
+			document.querySelectorAll(`#${passwordId}`).forEach((ele) =>{
+				ele.setAttribute('type', 'password');
+			});
 			$(element).addClass('fa-eye-slash');
 			$(element).removeClass('fa-eye');
 		} else if ($(`#${passwordId}`).attr('type') === 'password') {
-			$(`#${passwordId}`).attr('type', 'text');
+			document.querySelectorAll(`#${passwordId}`).forEach((ele) =>{
+				ele.setAttribute('type', 'text');
+			});
 			$(element).removeClass('fa-eye-slash');
 			$(element).addClass('fa-eye');
 		}